+++ indri/branches/strohman/include/indri/TermScoreFunction.hpp
@@ -26,8 +26,8 @@
  */
 class TermScoreFunction {
 public:
-  virtual double scoreOccurrence( int occurrences, int contextLength ) = 0;
-  virtual double scoreOccurrence( int occurrences, int contextLength, int documentOccurrences, int documentLength ) = 0;
+  virtual double scoreOccurrence( double occurrences, int contextLength ) = 0;
+  virtual double scoreOccurrence( double occurrences, int contextLength, double documentOccurrences, int documentLength ) = 0;
 };
 
 #endif // INDRI_TERMSCOREFUNCTION_HPP
